hireportai/hirelens-backend/requirements-lock.txt
aiofiles==25.1.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.91.0
anyio==4.13.0
asyncpg==0.31.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
confection==1.3.3
cryptography==46.0.7
cymem==2.0.13
distro==1.9.0
docstring_parser==0.17.0
ecdsa==0.19.2
fastapi==0.135.3
fsrs==6.3.1
google-auth==2.49.1
google-genai==1.70.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
numpy==2.4.4
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pgvector==0.4.2
pillow==12.2.0
pluggy==1.6.0
preshed==3.0.13
psycopg==3.3.3
psycopg-binary==3.3.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pypdfium2==5.6.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-docx==1.2.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.24
PyYAML==6.0.3
requests==2.33.1
rich==14.3.3
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.49
srsly==2.5.3
starlette==1.0.0
stripe==15.0.1
tenacity==9.1.4
thinc==8.3.13
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==1.0.0
websockets==16.0
wrapt==2.1.2
